+++ api/result/errors.go
@@ -2,4 +2,4 @@ package result
 
 import "github.com/pkg/errors"
 
-var ErrNotFoundField = errors.New("field for get data result not found")
+var ErrNotFoundField = errors.New("field for get data result not found") // "field for get data result not found"